Fix a bug

+++ fract/model/bet.py
@@ -22,7 +22,8 @@ class BettingSystem(object):
     def calculate_size_by_pl(self, unit_size, inst_pl_txns, init_size=None):
         pl_list = [float(t['pl']) for t in inst_pl_txns if float(t['pl']) != 0]
         size_list = [
-            abs(int(t['units'])) for t in inst_pl_txns if int(t['units']) != 0
+            abs(int(float(t['units']))) for t in inst_pl_txns
+            if float(t['units']) != 0
         ]
         if not (pl_list and size_list):
             return init_size or unit_size
